Move WarehouseInventoryAdjustmentAdvices schema to an older version namespace.

+++ generate.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 set -e
 cd "$(dirname "${BASH_SOURCE[0]}")"
-version=v777
+versions=(v710 v777)
 apis=(
 "CarrierFreightDetails.json|CarrierFreightDetail"
 "CarrierLoadTenderResponses.json|CarrierLoadTenderResponse"
@@ -20,17 +20,20 @@ apis=(
 "WarehouseInventoryAdjustmentAdvices.json|WarehouseInventoryAdjustmentAdvice"
 "WarehouseTransferReceiptAdvices.json|WarehouseTransferReceiptAdvice"
 )
-for api in "${apis[@]}"; do
-  file=${api%%|*}
-  class=${api##*|}
-  name=${file%.json}
-  echo "Generating $class classes for $file schema"
-  mkdir -p src/RSX/$version/$name
-  docker run --rm -v $PWD:/app --workdir /app \
-    swaggest/json-cli \
-    json-cli gen-php "json-schema/$version/$file" \
-      --ptr-in-schema "#/definitions/$class" \
-      --def-ptr "#/definitions" \
-      --ns ShipStream\\SpsCommerce\\RSX\\$version\\$name \
-      --ns-path src/RSX/$version/$name/
+for version in "${versions[@]}"; do
+  for api in "${apis[@]}"; do
+    file=${api%%|*}
+    class=${api##*|}
+    name=${file%.json}
+    [ ! -f "json-schema/$version/$file" ] && continue
+    echo "Generating $class classes for $version/$file schema"
+    mkdir -p src/RSX/$version/$name
+    docker run --rm -v $PWD:/app --workdir /app \
+      swaggest/json-cli \
+      json-cli gen-php "json-schema/$version/$file" \
+        --ptr-in-schema "#/definitions/$class" \
+        --def-ptr "#/definitions" \
+        --ns ShipStream\\SpsCommerce\\RSX\\$version\\$name \
+        --ns-path src/RSX/$version/$name/
+  done
 done
